iota_grpc_types::v1

Module filter

Modules

command_filter

Nested message and enum types in CommandFilter.

event_filter

Nested message and enum types in EventFilter.

transaction_filter

Nested message and enum types in TransactionFilter.

Structs

AddressFilter

Filter by address.

AddressFilterFieldPathBuilder

AllEventFilter

Logical AND of several filters.

AllEventFilterFieldPathBuilder

AllTransactionFilter

Logical AND of several filters.

AllTransactionFilterFieldPathBuilder

AnyEventFilter

Logical OR of several filters.

AnyEventFilterFieldPathBuilder

AnyTransactionFilter

Logical OR of several filters.

AnyTransactionFilterFieldPathBuilder

CommandFilter

Filter by command type.

CommandFilterFieldPathBuilder

EventFilter

Filter for events.

EventFilterFieldPathBuilder

ExecutionStatusFilter

Filter by transaction execution status. Set success to true to match successful transactions, or false to match failed transactions (cancelled, execution error, etc.).

ExecutionStatusFilterFieldPathBuilder

MakeMoveVecCommandFilter

Match a MakeMoveVec command.

MakeMoveVecCommandFilterFieldPathBuilder

MergeCoinsCommandFilter

Match a MergeCoins command.

MergeCoinsCommandFilterFieldPathBuilder

MoveCallCommandFilter

Filter by move package, module (optional) and function (optional).

MoveCallCommandFilterFieldPathBuilder

MoveEventTypeFilter

Filter by the given Move event struct name (struct tag). For example, if the event is defined in 0xabcd::MyModule, and named Foo, then the struct tag is 0xabcd::MyModule::Foo.

MoveEventTypeFilterFieldPathBuilder

MovePackageAndModuleFilter

Filter by Move package + module (optional).

MovePackageAndModuleFilterFieldPathBuilder

NotEventFilter

Logical NOT of a filter.

NotEventFilterFieldPathBuilder

NotTransactionFilter

Logical NOT of a filter.

NotTransactionFilterFieldPathBuilder

ObjectIdFilter

Filter by input object ID.

ObjectIdFilterFieldPathBuilder

PublishCommandFilter

Match a Publish command.

PublishCommandFilterFieldPathBuilder

SplitCoinsCommandFilter

Match a SplitCoins command.

SplitCoinsCommandFilterFieldPathBuilder

TransactionFilter

Filter for transactions.

TransactionFilterFieldPathBuilder

TransactionKindsFilter

Filter by transaction kinds (any of the specified kinds).

TransactionKindsFilterFieldPathBuilder

TransferObjectsCommandFilter

Match a TransferObjects command.

TransferObjectsCommandFilterFieldPathBuilder

UpgradeCommandFilter

Match an Upgrade command. Optionally filter by the specific package being upgraded.

UpgradeCommandFilterFieldPathBuilder

Enums

TransactionKind