iota_grpc_types::v1

Module transaction_execution_service

Modules

execute_transaction_result

Nested message and enum types in ExecuteTransactionResult.

simulate_transaction_item

Nested message and enum types in SimulateTransactionItem.

simulate_transaction_result

Nested message and enum types in SimulateTransactionResult.

simulated_transaction

Nested message and enum types in SimulatedTransaction.

transaction_execution_service_client

Generated client implementations.

transaction_execution_service_server

Generated server implementations.

Structs

ExecuteTransactionItem

A single transaction to execute, along with its authorizing signatures.

ExecuteTransactionItemFieldPathBuilder

ExecuteTransactionResult

The result of executing a single transaction: either the executed transaction or an error.

ExecuteTransactionResultFieldPathBuilder

ExecuteTransactionsRequest

ExecuteTransactionsRequestFieldPathBuilder

ExecuteTransactionsResponse

Response message for TransactionExecutionService.ExecuteTransactions.

ExecuteTransactionsResponseFieldPathBuilder

ExecutionError

ExecutionErrorFieldPathBuilder

SimulateTransactionItem

A single transaction to simulate, along with simulation options.

SimulateTransactionItemFieldPathBuilder

SimulateTransactionResult

The result of simulating a single transaction: either the simulated transaction or an error.

SimulateTransactionResultFieldPathBuilder

SimulateTransactionsRequest

SimulateTransactionsRequestFieldPathBuilder

SimulateTransactionsResponse

Response message for TransactionExecutionService.SimulateTransactions.

SimulateTransactionsResponseFieldPathBuilder

SimulatedTransaction

The result of simulating a single transaction.

SimulatedTransactionFieldPathBuilder